+++ modules/services/crashlogging/src/main/java/au/com/shiftyjelly/pocketcasts/crashlogging/CrashReportPermissionCheck.kt
@@ -1,0 +1,5 @@
+package au.com.shiftyjelly.pocketcasts.crashlogging
+
+fun interface CrashReportPermissionCheck {
+    operator fun invoke(): Boolean
+}