+++ modules/services/crashlogging/src/main/java/au/com/shiftyjelly/pocketcasts/crashlogging/ExceptionsFilter.kt
@@ -1,24 +1,9 @@
-package au.com.shiftyjelly.pocketcasts.utils
+package au.com.shiftyjelly.pocketcasts.crashlogging
 
-import io.sentry.Sentry
 import java.io.IOException
 import java.util.concurrent.CancellationException
-import javax.net.ssl.SSLException
 
-object SentryHelper {
-    const val GLOBAL_TAG_APP_PLATFORM = "app.platform"
-
-    fun recordException(throwable: Throwable) {
-        if (shouldIgnoreExceptions(throwable)) {
-            return
-        }
-        Sentry.captureException(throwable)
-    }
-
-    fun recordException(message: String, throwable: Throwable) {
-        recordException(Exception(message, throwable))
-    }
-
+internal object ExceptionsFilter {
     fun shouldIgnoreExceptions(throwable: Throwable): Boolean {
         if (shouldIgnoreException(throwable)) {
             return true
@@ -40,15 +25,8 @@ object SentryHelper {
             // ignore worker job cancels as they will retry
             throwable is CancellationException ||
                 // ignore exceptions such as SocketTimeoutException, SocketException or UnknownHostException, as with episode urls we don't control this
-                throwable is IOException ||
-                // ignore producer certificate exceptions
-                throwable is SSLException
+                // or certificate exceptions
+                throwable is IOException
             )
-    }
-
-    enum class AppPlatform(val value: String) {
-        MOBILE("mobile"),
-        AUTOMOTIVE("automotive"),
-        WEAR("wear"),
     }
 }