+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/opml/OpmlImportTask.kt
@@ -28,12 +28,21 @@ import au.com.shiftyjelly.pocketcasts.servers.refresh.RefreshServerManager
 import au.com.shiftyjelly.pocketcasts.utils.log.LogBuffer
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedInject
+import java.io.BufferedReader
+import java.io.BufferedWriter
+import java.io.File
+import java.io.FileInputStream
+import java.io.FileOutputStream
 import java.io.InputStream
+import java.io.InputStreamReader
+import java.io.OutputStreamWriter
 import java.io.StringReader
 import java.net.URL
 import java.util.Scanner
 import java.util.regex.Pattern
 import javax.xml.parsers.SAXParserFactory
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.asFlow
@@ -41,9 +50,11 @@ import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.mapNotNull
 import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.flow.toList
+import kotlinx.coroutines.launch
 import org.xml.sax.Attributes
 import org.xml.sax.InputSource
 import org.xml.sax.SAXException
+import org.xml.sax.SAXParseException
 import org.xml.sax.helpers.DefaultHandler
 import au.com.shiftyjelly.pocketcasts.images.R as IR
 import au.com.shiftyjelly.pocketcasts.localization.R as LR
@@ -165,8 +176,16 @@ class OpmlImportTask @AssistedInject constructor(
             }
             val numberProcessed = processFile(uri)
             trackProcessed(numberProcessed)
+            CoroutineScope(Dispatchers.Main).launch {
+                Toast.makeText(applicationContext, applicationContext.getString(LR.string.settings_import_opml_succeeded_message), Toast.LENGTH_LONG).show()
+            }
             return Result.success()
         } catch (t: Throwable) {
+            if (t is SAXParseException) {
+                CoroutineScope(Dispatchers.Main).launch {
+                    Toast.makeText(applicationContext, applicationContext.getString(LR.string.settings_import_opml_import_failed_message), Toast.LENGTH_LONG).show()
+                }
+            }
             LogBuffer.e(LogBuffer.TAG_BACKGROUND_TASKS, t, "OPML import failed.")
             trackFailure(reason = "unknown")
             return Result.failure()
@@ -180,7 +199,7 @@ class OpmlImportTask @AssistedInject constructor(
         )
     }
 
-    fun trackFailure(reason: String) {
+    private fun trackFailure(reason: String) {
         analyticsTracker.track(
             AnalyticsEvent.OPML_IMPORT_FAILED,
             mapOf("reason" to reason),
@@ -216,7 +235,8 @@ class OpmlImportTask @AssistedInject constructor(
         val resolver = applicationContext.contentResolver
         try {
             resolver.openInputStream(uri)?.use { inputStream ->
-                urls = readOpmlUrlsSax(inputStream)
+                val modifiedInputStream = replaceInvalidXmlCharacter(inputStream)
+                urls = readOpmlUrlsSax(modifiedInputStream)
             }
         } catch (e: SAXException) {
             resolver.openInputStream(uri)?.use { inputStream ->
@@ -226,6 +246,37 @@ class OpmlImportTask @AssistedInject constructor(
 
         processUrls(urls)
         return urls.size
+    }
+
+    private fun replaceInvalidXmlCharacter(inputStream: InputStream): InputStream {
+        val tempFile = File.createTempFile("output", ".xml")
+        tempFile.deleteOnExit()
+
+        BufferedReader(InputStreamReader(inputStream, Charsets.UTF_8)).use { reader ->
+            BufferedWriter(OutputStreamWriter(FileOutputStream(tempFile), Charsets.UTF_8)).use { writer ->
+                reader.lineSequence().forEach { line ->
+                    val modifiedLine = line
+                        .replace("<outline", "\n<outline")
+                        .replace("&(?!amp;|quot;|gt;|lt;)".toRegex(), "&amp;")
+                        .replace("’", "&apos;")
+
+                    val textRegex = """(?<=text=")(.*?)(?="\s*(?:/>|\s+xmlUrl|\s+type))""".toRegex()
+                    val finalLine = textRegex.replace(modifiedLine) { matchResult ->
+                        val text = matchResult.groupValues[1]
+                        val fixedText = text
+                            .replace("\"", "&quot;")
+                            .replace(">", "&gt;")
+                            .replace("<", "&lt;")
+                        matchResult.value.replace(text, fixedText)
+                    }
+
+                    writer.write(finalLine)
+                    writer.newLine()
+                }
+            }
+        }
+
+        return FileInputStream(tempFile)
     }
 
     private suspend fun processUrls(urls: List<String>) {