+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/playback/auto/AutoConverter.kt
@@ -1,6 +1,8 @@
 package au.com.shiftyjelly.pocketcasts.repositories.playback.auto
 
+import android.content.ContentResolver
 import android.content.Context
+import android.graphics.Bitmap
 import android.graphics.Color
 import android.net.Uri
 import android.os.Bundle
@@ -11,6 +13,7 @@ import android.support.v4.media.MediaDescriptionCompat.STATUS_DOWNLOADED
 import android.support.v4.media.MediaDescriptionCompat.STATUS_NOT_DOWNLOADED
 import android.util.Log
 import androidx.annotation.DrawableRes
+import androidx.core.graphics.drawable.toBitmap
 import androidx.core.os.bundleOf
 import androidx.media.utils.MediaConstants.DESCRIPTION_EXTRAS_KEY_COMPLETION_STATUS
 import androidx.media.utils.MediaConstants.DESCRIPTION_EXTRAS_VALUE_COMPLETION_STATUS_FULLY_PLAYED
@@ -29,9 +32,12 @@ import au.com.shiftyjelly.pocketcasts.repositories.extensions.autoDrawableId
 import au.com.shiftyjelly.pocketcasts.repositories.extensions.automotiveDrawableId
 import au.com.shiftyjelly.pocketcasts.repositories.extensions.getArtworkUrl
 import au.com.shiftyjelly.pocketcasts.repositories.extensions.getSummaryText
+import au.com.shiftyjelly.pocketcasts.repositories.images.PocketCastsImageRequestFactory
 import au.com.shiftyjelly.pocketcasts.repositories.playback.EXTRA_CONTENT_STYLE_GROUP_TITLE_HINT
 import au.com.shiftyjelly.pocketcasts.repositories.playback.FOLDER_ROOT_PREFIX
 import au.com.shiftyjelly.pocketcasts.utils.Util
+import coil.executeBlocking
+import coil.imageLoader
 import com.bumptech.glide.load.DataSource
 import com.bumptech.glide.load.engine.GlideException
 import com.bumptech.glide.request.RequestListener
@@ -63,7 +69,7 @@ data class AutoMediaId(
 
 object AutoConverter {
     fun convertEpisodeToMediaItem(context: Context, episode: BaseEpisode, parentPodcast: Podcast, useEpisodeArtwork: Boolean, groupTrailers: Boolean = false, sourceId: String = parentPodcast.uuid): MediaBrowserCompat.MediaItem {
-        val localUri = getBitmapUriForPodcast(parentPodcast, episode, context, useEpisodeArtwork)
+        val localUri = getPodcastArtworkUri(parentPodcast, episode, context, useEpisodeArtwork)
 
         val extrasForEpisode = extrasForEpisode(episode)
         if (groupTrailers) {
@@ -85,7 +91,7 @@ object AutoConverter {
 
     fun convertPodcastToMediaItem(podcast: Podcast, context: Context, useEpisodeArtwork: Boolean): MediaBrowserCompat.MediaItem? {
         return try {
-            val localUri = getBitmapUriForPodcast(podcast = podcast, episode = null, context = context, showRssArtwork = useEpisodeArtwork)
+            val localUri = getPodcastArtworkUri(podcast = podcast, episode = null, context = context, showRssArtwork = useEpisodeArtwork)
 
             val podcastDesc = MediaDescriptionCompat.Builder()
                 .setTitle(podcast.title)
@@ -125,7 +131,7 @@ object AutoConverter {
         return MediaBrowserCompat.MediaItem(mediaDescription, MediaBrowserCompat.MediaItem.FLAG_BROWSABLE)
     }
 
-    fun getBitmapUriForPodcast(podcast: Podcast?, episode: BaseEpisode?, context: Context, showRssArtwork: Boolean): Uri? {
+    fun getPodcastArtworkUri(podcast: Podcast?, episode: BaseEpisode?, context: Context, showRssArtwork: Boolean): Uri? {
         val url = if (episode is PodcastEpisode && (!episode.imageUrl.isNullOrBlank()) && showRssArtwork) {
             episode.imageUrl
         } else if (episode is UserEpisode) {
@@ -141,6 +147,23 @@ object AutoConverter {
 
         val podcastArtUri = Uri.parse(url)
         return getArtworkUriForContentProvider(podcastArtUri, context)
+    }
+
+    fun getPodcastArtworkBitmap(episode: BaseEpisode, context: Context, useEpisodeArtwork: Boolean): Bitmap? {
+        val imageRequestFactory = PocketCastsImageRequestFactory(
+            context,
+            isDarkTheme = true,
+            size = 480,
+            placeholderType = PocketCastsImageRequestFactory.PlaceholderType.Small,
+        )
+
+        val request = imageRequestFactory.create(episode, useEpisodeArtwork)
+        return context.imageLoader.executeBlocking(request).drawable?.toBitmap() ?: loadPlaceholderBitmap(imageRequestFactory, context)
+    }
+
+    private fun loadPlaceholderBitmap(imageRequestFactory: PocketCastsImageRequestFactory, context: Context): Bitmap? {
+        val request = imageRequestFactory.createForPodcast(podcastUuid = null)
+        return context.imageLoader.executeBlocking(request).drawable?.toBitmap()
     }
 
     private fun getBitmapUriForFolder(context: Context, folder: Folder?): Uri? {
@@ -194,8 +217,14 @@ object AutoConverter {
      * Use the drawable id so Proguard doesn't remove the asset in the production build.
      */
     fun getBitmapUri(@DrawableRes drawable: Int, context: Context): Uri {
-        val drawableName = context.resources.getResourceEntryName(drawable)
-        return Uri.parse("android.resource://" + context.packageName + "/drawable/" + drawableName)
+        val resources = context.resources
+        // This is an example of the URI android.resource://au.com.shiftyjelly.pocketcasts/drawable/auto_folder_01
+        return Uri.Builder()
+            .scheme(ContentResolver.SCHEME_ANDROID_RESOURCE) // android.resource
+            .authority(resources.getResourcePackageName(drawable)) // au.com.shiftyjelly.pocketcasts
+            .appendPath(resources.getResourceTypeName(drawable)) // drawable
+            .appendPath(resources.getResourceEntryName(drawable)) // auto_folder_01
+            .build()
     }
 
     private fun extrasForEpisode(episode: BaseEpisode): Bundle {