+++ modules/services/repositories/src/test/java/au/com/shiftyjelly/pocketcasts/repositories/podcast/EpisodeManagerImplTest.kt
@@ -1,12 +1,15 @@
 package au.com.shiftyjelly.pocketcasts.repositories.podcast
 
+import app.cash.turbine.test
 import au.com.shiftyjelly.pocketcasts.models.db.AppDatabase
 import au.com.shiftyjelly.pocketcasts.models.db.dao.EpisodeDao
 import au.com.shiftyjelly.pocketcasts.models.entity.ChapterIndices
 import au.com.shiftyjelly.pocketcasts.models.entity.PodcastEpisode
 import au.com.shiftyjelly.pocketcasts.sharedtest.MainCoroutineRule
+import java.util.Date
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
+import org.junit.Assert.assertEquals
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
@@ -15,8 +18,10 @@ import org.mockito.Mock
 import org.mockito.Mockito.verify
 import org.mockito.junit.MockitoJUnitRunner
 import org.mockito.kotlin.any
+import org.mockito.kotlin.doReturnConsecutively
 import org.mockito.kotlin.mock
 import org.mockito.kotlin.never
+import org.mockito.kotlin.stub
 import org.mockito.kotlin.whenever
 
 @OptIn(ExperimentalCoroutinesApi::class)
@@ -77,5 +82,25 @@ class EpisodeManagerImplTest {
         episodeManagerImpl.deselectChapterIndexForEpisode(3, episode)
 
         verify(episodeDao, never()).update(any())
+    }
+
+    @Test
+    fun `get all podcasts episodes`() = runTest {
+        val episodes = List(26) { PodcastEpisode(uuid = "$it", publishedDate = Date()) }
+        episodeDao.stub {
+            onBlocking { getAllPodcastEpisodes(any(), any()) } doReturnConsecutively (episodes.chunked(10) + listOf(emptyList()))
+        }
+
+        episodeManagerImpl.getAllPodcastEpisodes(10).test {
+            episodes.forEachIndexed { index, episode ->
+                assertEquals(episode to index, awaitItem())
+            }
+
+            expectNoEvents()
+        }
+
+        verify(episodeDao).getAllPodcastEpisodes(10, 0)
+        verify(episodeDao).getAllPodcastEpisodes(10, 10)
+        verify(episodeDao).getAllPodcastEpisodes(10, 20)
     }
 }